fixed access control issue

+++ index.ts
@@ -5,13 +5,13 @@ import cors = require("cors");
 import jwt = require("express-jwt");
 import jwksRsa = require("jwks-rsa");
 import { getUserByUserID } from "./directory";
-import { initDb, getTodos, insertTodo, updateTodo, deleteTodo } from "./store";
+import { initDb, getTodos, getTodo, insertTodo, updateTodo, deleteTodo } from "./store";
 import { UserCache, User } from "./interfaces";
 import * as dotenv from "dotenv";
 
 dotenv.config();
 
-import { jwtAuthz } from "@aserto/aserto-node";
+import { jwtAuthz, is } from "@aserto/aserto-node";
 import { getConfig } from "./config";
 
 const authzOptions = getConfig();
@@ -42,6 +42,8 @@ const PORT = 3001;
 
 //Users cache
 const users: UserCache = {};
+
+// use checkAuthz middleware to authorize the operation
 app.get("/user/:userID", checkJwt, checkAuthz, async (req, res) => {
   const { userID } = req.params;
   const user: User = users[userID]
@@ -53,6 +55,7 @@ app.get("/user/:userID", checkJwt, checkAuthz, async (req, res) => {
   res.json(user);
 });
 
+// use checkAuthz middleware to authorize the operation
 app.get("/todos", checkJwt, checkAuthz, async (req, res) => {
   try {
     const todos: Todo[] = await getTodos();
@@ -62,6 +65,7 @@ app.get("/todos", checkJwt, checkAuthz, async (req, res) => {
   }
 });
 
+// use checkAuthz middleware to authorize the operation
 app.post("/todo", checkJwt, checkAuthz, async (req, res) => {
   const todo: Todo = req.body;
   try {
@@ -72,21 +76,41 @@ app.post("/todo", checkJwt, checkAuthz, async (req, res) => {
   }
 });
 
-app.put("/todo/:ownerID", checkJwt, checkAuthz, async (req, res) => {
-  const todo: Todo = req.body;
+app.put("/todo/:ownerID", checkJwt, async (req, res) => {
+  const body: Todo = req.body;
   try {
-    await updateTodo(todo);
-    res.json({ msg: "Todo updated" });
+    // retrieve todo from db, and overwrite updateable properties with incoming payload
+    const todo = await getTodo(body);
+    todo.Completed = body.Completed;
+    todo.Title = body.Title;
+
+    // call authorizer to determine whether operation is allowed
+    // resource context is passed in as the OwnerID of the todo in the database
+    if (await is("allowed", req, authzOptions, null, { ownerID: todo.OwnerID })) {
+      await updateTodo(todo);
+      res.json({ msg: "Todo updated" });
+    } else {
+      res.status(403).send({ "msg": "not allowed" })
+    }
   } catch (error) {
     res.status(500).send(error);
   }
 });
 
-app.delete("/todo/:ownerID", checkJwt, checkAuthz, async (req, res) => {
-  const todo: Todo = req.body;
+app.delete("/todo/:ownerID", checkJwt, async (req, res) => {
+  const body: Todo = req.body;
   try {
-    deleteTodo(todo);
-    res.json({ msg: "Todo deleted" });
+    // retrieve todo from db via ID in incoming payload
+    const todo = await getTodo(body);
+
+    // call authorizer to determine whether operation is allowed
+    // resource context is passed in as the OwnerID of the todo in the database
+    if (await is("allowed", req, authzOptions, null, { ownerID: todo.OwnerID })) {
+      await deleteTodo(todo);
+      res.json({ msg: "Todo deleted" });
+    } else {
+      res.status(403).send({ "message": "not allowed" })
+    }
   } catch (e) {
     res.status(500).send(e);
   }
